python-backend/requirements.txt
flask
flask_cors
oracledb
python-dotenv